### Jira ticket url
- https://zuelligpharma.atlassian.net/browse/EZRX-19880

### Changes
- Add market place logo and seller name if the item is from MarketPlace
- Change color of textfield to match design
- Limit number of character of Request counter offer input
- Show confirmation popup when deleting attachments

### Test account
- environment: mock
- username: rootadmin
- salesOrg: HK
- customer: default

### Output screenshot
-![Screenshot 2024-03-26 at 09.25.29.png](https://dev.azure.com/zuelligpharmadevops/3368559d-402f-4ca8-a029-737bca870679/_apis/git/repositories/d10de74c-8ac2-4e8e-af4b-a023b6c9dee2/pullRequests/99591/attachments/Screenshot%202024-03-26%20at%2009.25.29.png)
![Screenshot 2024-03-27 at 18.36.32.png](https://dev.azure.com/zuelligpharmadevops/3368559d-402f-4ca8-a029-737bca870679/_apis/git/repositories/d10de74c-8ac2-4e8e-af4b-a023b6c9dee2/pullRequests/99591/attachments/Screenshot%202024-03-27%20at%2018.36.32.png)

+++ lib/presentation/returns/new_request/tabs/return_details_tab/widgets/upload_attachement_section.dart
@@ -228,19 +228,23 @@ class _UploadedFileList extends StatelessWidget {
                           IconButton(
                             icon: const Icon(Icons.delete_outline_outlined),
                             padding: const EdgeInsets.all(0.0),
-                            onPressed: () {
-                              context.read<ReturnRequestAttachmentBloc>().add(
-                                    ReturnRequestAttachmentEvent.deleteFile(
-                                      file: file,
-                                    ),
-                                  );
-                              context.read<NewRequestBloc>().add(
-                                    NewRequestEvent.toggleFiles(
-                                      files: [file],
-                                      included: false,
-                                      uuid: data.uuid,
-                                    ),
-                                  );
+                            onPressed: () async {
+                              final confirmed =
+                                  await _showConfirmBottomSheet(context);
+                              if ((confirmed ?? false) && context.mounted) {
+                                context.read<ReturnRequestAttachmentBloc>().add(
+                                      ReturnRequestAttachmentEvent.deleteFile(
+                                        file: file,
+                                      ),
+                                    );
+                                context.read<NewRequestBloc>().add(
+                                      NewRequestEvent.toggleFiles(
+                                        files: [file],
+                                        included: false,
+                                        uuid: data.uuid,
+                                      ),
+                                    );
+                              }
                             },
                           ),
                         ],
@@ -251,5 +255,18 @@ class _UploadedFileList extends StatelessWidget {
             ],
           )
         : const SizedBox.shrink();
+  }
+
+  Future<bool?> _showConfirmBottomSheet(BuildContext context) {
+    return showModalBottomSheet<bool>(
+      context: context,
+      isScrollControlled: true,
+      enableDrag: false,
+      builder: (_) => const ConfirmBottomSheet(
+        title: 'Remove item?',
+        content: 'This action cannot be undone',
+        confirmButtonText: 'Remove',
+      ),
+    );
   }
 }
